requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.2
idna==2.8
multidict==4.5.2
python-dateutil==2.8.0
schedule==0.6.0
six==1.12.0
websockets==6.0
yarl==1.3.0
